Use Telemetry for Webcore response metrics
This updates `Services.Webcore` to use `Belfrage.Metrics` instead of
`Belfrage.Event` to track metrics related to Webcore responses. It also
adds a config to convert the new telemetry events into metrics (the
metrics themselves remained the same).

`Belfrage.Metrics.event` can now accept metadata (this was necessary so
that webcore status / error codes could be passed as metadata) and
`assert_metrics` test helper can now accept just a name or a `{name,
metadata} tuple as a metric. The latter means that it can be used
instead of `assert_event`, so I removed it so that we have a single test
helper for all metrics.