Add Belfrage.AWS.HttpClient
Before this change `Belfrage.Clients.Lambda` had two purposes:

* It was a wrapper for making calls to AWS Lambda that would call
`ExAws` and interpret the response.
* It was an HTTP client that `ExAws` was configured to use for all
requests to  AWS (not just AWS Lambda).

The latter was confusing: why something that is called
`Belfrage.Clients.Lambda` is used to make calls to AWS STS for example.
It was also quite rigid because things like the request timeout and the
name of the HTTP workers pool was hardcoded. This meant that we used the
timeout configured for AWS Lambda calls for any AWS calls and all calls
to AWS used the `:Webcore` HTTP pool.

We currently only use `ExAWS` to make calls to Pres lambda and to AWS
STS to get credentials for that call, so this didn't have a big
practical effect, and I'm making this change to reduce possible
confusion.