Use Telemetry to track duration of calls to Pres lambda
We'll use `Belfrage.Metrics.duration` instead of `Belfrage.Event.record`
to track the duration of calls to Pres lambda because the former relies
on telemetry and convert the emitted event to a metric.

We'll do it in `Services.Webcore` so that we could later add metadata
(like the name of the route spec) to the event.