Use Telemetry for plug metrics
We'll need to tag the metrics that track the number of requests Belfrage
receives and responses it returnes with route_spec so that other teams
could filter them. Other teams might be interested in these metrics in
addition to platform-level metrics because Belfrage returns cached
responses and also errors can happen in Belfrage itself.

I thought I'd start using Telemetry for these metrics to make it easier
to add dimensions to them. `Plug` ships with a `Telemetry` plug that
does pretty much what our our custom plug used to do: emits events
start/stop events with a duration and `Plug.Conn` as metadata. We can
use those events to generate the metrics.

I thought that we should put `Plug.Telemetry` to the very beginning of
our plug pipeline so that we emit events before any redirects, and
include as much as possible in the duration measured.