Add new plug metrics with route_spec as dimension
This adds the following new metrics:

* `belfrage.request.duration` - duration of request to Belfrage with
`:status_code` and `route_spec` as dimensions
* `belfrage.response` - counter of responses from Belfrage with
`status_code` and `route_spec` as dimensions
* `belfrage.response.private` - counter of responses with
'cache-control: private' header from Belfrage with `status_code` and
`route_spec` as dimensions
* `belfrage.response.stale` - counter of stale cached response from
Belfrage with `route_spec` as dimension
* `belfrage.error` - counter of errors in Belfrage with `route_spec` as
dimension

These new metrics should make it easier for other teams to filter
request/response metrics by route spec and therefore enable capability
teams to see metrics that are relevant to them.

I thought I'd add some new metrics to

* Make them more convenient to use: existing metrics like
`web.response.status.X` have status codes in the name which means that
one needs to specify them when querying.
* Provide data that was misssing (e.g. number of requests that fail
because of an error in Belfrage).
* Ensure the naming of metrics is concise and consistent.

These new metrics are similar to Webcore-specific metrics added in
6bdc4b4670af9b6f8243878e7f6a5c8710835038

To be able to report route spec on internal errors in Belfrage I had to
store it in `Plug.Conn.assigns`, catch any errors in
`RouteMaster.yield/2` and wrap them in a `Plug.Conn.WrapperError` which
allows to set the `conn` and therefore preserve the assigned route spec.
If we don't do this, then the `conn` that is used to render the response
won't contain `route_spec` in assigns and we won't be able to use it in
metrics.