Update Plug to version 1.12.1
I needed to update `Plug` to the most recent version because it changes
the way telemetry is used in `Plug.Router`: the name of the router
module is now included in metadata of `plug.router_dispatch.exception`
event which is helpful when there is a router module that delegates
routing to another module (in our case `BelfrageWeb.Router` delegates to
`BelfrageWeb.RoutefilePointer`).

There are no breaking changes in this new version apart from the removal
of `Plug.Conn.before_send` attribute which a couple of tests relied on.
I don't think that relying on that attribute was a good way to test that
our plugs do something after sending a response, so I updated those
tests (they emulate sending a response now).