Operate on RouteSpec struct after merging spec maps
Once route and platform specs are merged and we have a `RouteSpec`
struct, we should operate on it like it's a `RouteSpec` struct, not a
map. E.g. we probably shouldn't use `Map.put/3` on it because it may end
up an attribute that's not in defined in the struct, like `loop_id`.