MORPH-2808 add extra webcore TTL dials and update tested max-age to 5s (#1138)
We're planning to update the max-age in presentation layer to 5s to reflect the new default mode. However, this leaves Belfrage without many longer options that would be effective in an incident scenario - the largest multiplier is x4 which leaves it at 20s. 

We'd rather have up to 120s as an option, so we'll add more options and update the testing scenarios for webcore around this.