Refactor `RouteSpec.specs_for/2` (#1151)
When `RouteSpec.specs_for` returns a `RouteSpec` by name, it does the
following:

* Gets route attribute by calling `Routes.Specs.Name.specs/{0,1}`
function
* Gets the name of the platform from the route attributes and calls
`Routes.Platforms.PlatformName.specs/{0,1}` function to get platform
attributes.
* Uses platform attributes as a base and overwrites them with route
attributes, except allowlists and pipeline.
* Merges platform and route allowlists, but if platform allowlist is a
wildcard ("*"), it keeps that wildcard and discrards the route
allowlist.
* If the platform pipeline contains a placeholder, it replaces the
placeholder with the route pipeline. Otherwise it uses the route
pipeline.

The logic above was implemented by using `Map.merge/3` with a function
that was called when both platform and route specs contained the same
key (`merge_key`). It was a clever generic solution, but I think it
obscured the logic above a little bit, so I tried to refactor it to make
the logic above more explicit. I wanted to make it clearer that route
attributes always take priority except for allowlists and pipeline.

I also re-wrote the test to only call `RouteSpec.spec_for/1` function
and therefore test the final result (and not intermediate logic as
before).