Fix specifying wildcard in route allowlists
When wildcard is used in route allowlists, it should allow all values.
In other words if either the platform or route allowlist is a wildcard,
then the resulting allowlist should be a wildcard.

Before this change we'd attempt to add `"*"` to the platform allowlist
which resulted in a broken allowlist and caused errors down the line.