Don't check that Belfrage wasn't called in RouteMasterTest
It should be enough to check the response status / body / headers for
things like redirects and route validation.

If a request is passed on to Belfrage, it'll end up as a call to one of
the origins and that will fail because all origins are mocked anyway, so
it shouldn't be necessary to check that `Belfrage.handle` hasn't been
called.