Stub origins in RouteMasterTest
Stubbing origins should be enough to test that routing has worked: if it
has then the request will be made to the origin (and a successful
response will be returned back when the origin is stubbed). If it
hasn't, than an error would be returned.

The ID of the route spec is now stored in `Plug.Conn`, so could be
tested that way if necessary.

Stubbing origins is a more universal way than mocking `Belfrage` module
that could be used in any tests.