Use the same origin URLs in all test envs
The URLs themselves don't matter in tests, so using the same ones allows
to unify the configuration for the environments.

Also updated `AllowAllQsOverrideTest` to always set
`production_environment` back to `"test"`, even if it fails. Before this
change if that test failed, the value remained `"live"` and caused
confusing failures in other tests.