Remove end-to-end Mix env
Now when there are no differences between the `test` and `end_to_end`
Mix environments, we can use the `test` env to run all tests.

Using the same environment is better because code doesn't need to be
compiled multiple times when running different types of tests and it
makes it easier to run all tests using a single command.

I renamed the extension of end-to-end tests from `.ex` to `.exs` so that
the file matcher pattern that is used by `mix test` task sees them.
Using a `.ex` extension was a hack anyway: all Elixir tests are supposed
to have `.exs` extension.