Rename BelfrageWeb.View to BelfrageWeb.Response
`View` was not an appropriate name for the module: it's not a 'view'
part of an MVC architecture for example.

It's a module that effectively converts a `Struct` back to `Plug.Conn`:
it transforms the result of processing a request in Belfrage to a
response to send back to the client. I therefore thought that `Response`
would be a better name.