Rename BelfrageWeb.Response.render to put
I thought that `render` wasn't a good name for the function because it
doesn't "render" anything, i.e. it doesn't e.g. take a template and
transforms it to HTML like web frameworks do. It sets the response body
and headers on the passed `Plug.Conn` with values from the passed
`Struct`. I therefore thought that `put` would be a better name: as in
'put the response from struct into conn`

I also modified the order of arguments in `put` and `redirect` functions
to make `Plug.Conn` the first argument because the function is 'putting'
the response into it.