Preserve %Struct{} data when errors happen in Belfrage
This change was prompted by the need for better handling of internal
errors in Belfrage that happen when processing personalised requests.
Error responses generated in such case should be private.

Before this change if an error happened at some point of executing
`Belfrage.handle/1`, it would be intercepted by `Plug.ErrorHandler`
which would then call `BelfrageWeb.Response` to generate an error
response. The `%Struct{}` wouldn't be available at that point and so it
would not be possible to tell if the request was personalised or not.
This means that the generated error response would be public (which is
the default for internal Belfrage errors).

We now add the current `%Struct{}` to errors that happen during main
operations executed by `Belfrage.handle` to make it available to
`BelfrageWeb.Response` when it generates an error response. It means
that if an error happens after we've determined that the request is
personalised, that information will no longer be lost and will be
available when generating an error response.

Errors can happen at any time and e.g. before routing or personalisation
logic is executed and in those cases we might still return a public
response to a personalised request, so this doesn't completely fix the
issue, but it's probably the best thing we can do.

Also, in this commit we effectively re-implement something that `Plug`
is able to do: calling bits of logic (plugs) that operate on a token
state (conn) and wrapping those calls in a `try/catch`
(Plug.ErrorHandler). We do the same thing: call functions/pipelines that
operate on `%Struct{}` and wrap those calls in a `try/catch`. We also
have to wrap/re-wrap errors in `RouteMaster` because that's the
integration point between `Plug` and `Belfrage`.

I think this could be reason to consider whether drawing a hard line
between `Plug` and `Belfrage` is paying off: maybe `Belfrage` could be
implemented as a number of plugs instead of being a completely separate
layer that has to re-implement the behaviour of `Plug`.