Pass routefile to Router in route tests
This commit makes it possible to pass a routefile as an option when
calling the `Router` plug and uses that to specify the router in route
tests instead of relying on `RoutefilePointer` to set it.

This will allow to remove the `:routes_test` Mix environment and use the
`:test` env instead: the routefile is the only difference between the
two.