Changed tests to mock circuit breaker dial value
This has to be done because in the previous version of the circuit breaker transformer, there are two branching statements.
One is for if the threshold has been exceeded, the other is to query the dial state.
Originally in these tests the threshold is never exceeded so the dial state is never queried.

Now because the state of the dial is passed to another function (Belfrage.CircuitBreaker.apply_circuit_breaker?/2) the value of the dial has to be queried before that function can run.
The result of this is that for these test we have to mock the dial where we didn't before.