[RESFRAME-4509] Add next_throughput/2
As part of the work to improve the circuit breaker, we need to ensure
that we have the ability to increment the request throughput to an
origin - as long as the error count is not greater than a given threshold.
When the maximum throughput is reached, the throughput should not be
incremented.

If the error count does exceed the threshold, then we should set the
request throughput to 0.

Given this, the next throughput should be calculated as follows:

    | Error count exceeded threshold ? | Throughput | Next throughput |
    |----------------------------------|------------|-----------------|
    | true                             | -          | 0               |
    | false                            | 100        | 100             |
    | false                            | 0          | 10              |
    | false                            | 10         | 20              |
    | false                            | 20         | 60              |
    | false                            | 60         | 100             |