Use throughput to determine if circuit breaker is applied to request
`CircuitBreaker.apply?/2` previously checked if the error count
exceeded a given threshold - this is now the concern of
`CircuitBreaker.next_throughput/2`, and we can check
`struct.private.throughput` to determine whether or not to apply
the circuit breaker. We do this by introducing a function called
`sampled?/1` which takes a struct as an argument. `sampled?/1`
returns `true` if a random number between 1 and 100 is less than
or equal to `struct.private.throughput` and `false` otherwise.

The relevant tests have also been updated to use throughput values
rather than error count values, and any redundant tests have been
removed.