[RESFRAME-4576] Append allowlists when processing requests
Before this change, the header and cookie allowlists were
appended at the route spec level, using the function
`Belfrage.Personalisation.transform_route_spec/1`.

This does not work if properties at the request/struct level need to
be considered when these allowlists are appended. In this case,
the appending logic needs to moved to the request processing level.

The appending logic now takes place in
`Belfrage.Processor.allowlists/1`. Before the headers are filtered
using the allowlists, we append the relevant ones accordingly.