Adjusted the infinite loop guardian to accomodate Joan
We need to do this as we are introducing the Joan Belfrage stack to sit infront
of Mozart.  Sometimes Mozart routes requests it can't handle to Bruce. This
would make the `req-svc-chain` look like this `GTM,BELFRAGE,MOZART,BELFRAGE`.
With our current implementation of the infinite loop guardian a now legitimate
request would trigger it.

This work changes the infinite loop guardian so that it will only stop a
request if it sees Belfrage **more than 2 times** in the `req-svc-chain`. This
will mean that the request that are infinitely looping will be caught a cycle
later than previously, but legitimate requests will be allowed.