[RESFRAME-4581] Increase scope of infinite loop checks
As per this comment:
https://github.com/bbc/belfrage/pull/1260#discussion_r831890435 we can
see the current logic would not cover all cases because we have language
variants for news pages and other pages which could 'InfiniteLoop' that
are not limited to the '/news' scope.
The changes made here are to open the scope up of the single occurence
of the Belfrage in the req-svc-chain as the previous logic didn't really
care for the stack or specific routes. With this a second change was
also needed because we needed to allow Bruce stack loops not trigger the
InfiniteLoop detection on its first occurence of Belfrage in its
req-svc-chain.