Use Strings.starts_with? to check if news route
As per this suggestion here
https://github.com/bbc/belfrage/pull/1260/commits/000f24d8e8751c51d183c146cd1d1873dd20d5ae#r831892056
I used String.starts_with? instead of String.match? as its a much
simpler approach. In addition to this I added some test cases to cover
https://github.com/bbc/belfrage/pull/1260/commits/000f24d8e8751c51d183c146cd1d1873dd20d5ae#r831894404
to ensure we weren't introducing any bugs.

I attempted to look into :binary.match but was receiving some strange
errors so will investigate this further and potentially introduce it in
a seperate commit.