Introduces a robust parser for the `X-Amzn-Trace-Id` header
The trace header parser used in aws_ex_ray will cause an exception if the
header is malformed in any way. This is not something that we want as any
malformed header would result in a 500. This commit introduces a new parser
which will catch all errors in the header.