Introduces `Xray.parse_header/2` to `Plugs.Xray`
Originally a new trace would be started for every request in belfrage. Now we
are introducing the Joan stack, a belfrage stack might not be the first xray
instrumented service that an xray trace passes through. (Joan could route to
Mozart then Bruce). So we should only start a new trace if we don't receive the
`x-amzn-trace-id` from an upstream service. If we don't receive that header or
its malformed, we should start our own new trace.

This is what has been implemented in `Plugs.Xray`. A check is made to see if
the header `x-amzn-trace-id` exists, if its valid, a new segment is made using
the data within the header. Otherwise a new segment is made using a fresh
trace.