[RESFRAME-4607] Add personalised_route to cached response
When we cache a response and then fetch a fresh version of it
early, we skip the request pipeline in `Processor.request_pipeline/1`.

This means that `struct.private.personalised_route` will have the
original value and not the value updated by any transformers
in the `Processor.request_pipeline/1` call.

To remedy this, we store the `personalised_route` value in the
cached response when it is stored, and update
`struct.private.personalised_route` when it is fetched if it has a
value of `true`.