Adds logic for whether to use existing trace or start a new one.
The logic should go thusly.

If the xray header doesn't exist start a new one

If a request is from belfrage or made using local development, an attempt to
parse the xray header should always be made.

If a request isn't from belfrage or locally (from GTM or FASTLY) then we should
only parse that header if it looks like its come from a load balancer,
something in this PR I've been calling a 'partial' header for short hand. These
headers can be identifies as they only have the 'Root' key or the 'Root' and
'Self' key present. With these partial headers we should remove the 'Self' key
if its present, and then make a sampling decision and add the 'Sampled' key.

If request isn't from belfrage, and is a full (looks like its come from another
application using xray) new trace will be started rather than parsed.

The logic is like this so we can protect ourselves from people sending
malicious headers to get us to sample every trace. While also allowed load
balancing tracing to work, and of course for Bruce to be able to use Joan's
traces.