[RESFRAME-4650] Allow updates to new :mvt_seen field in RouteState
We would like to keep track of what pages are running mvt experiments.
We can do this by inspecting the vary header in the response returned
by downstream origins. If an experiment is being run against a page,
then the vary header will contain an mvt header. This header should
have the prefix 'mvt-'. We can then store it in the RouteState state
to keep track of the experiment across the pages/routes that use the
routespec.

RouteState.update/1 is similar to RouteState.inc/1 but also sends
the mvt vary headers in the message, when the response is not a
fallback and mvt vary headers exist in the vary header (the vary
header is in the response). Otherwise it just calls RouteState.inc/1.

When the message is received, the RouteState state is updated as follows:
The :mvt_seen field is updated with key-value pairs consisting of
the mvt vary header name and the current datetime respectively.

If the key already exists then the value is updated with the latest datetime.

This can be included in the response pipeline by replacing the
inc_route_state/1 call with an update_route_state_call/1:

```elixir
  defp update_route_state(struct = %Struct{}) do
    RouteState.update(struct)
    struct
  end
```