Place test setup in Belfrage.RouteStateTest in test/1 or describe/1
Previously there was a setup/0 block at the top of this test
module. The context then applies to all test and describe blocks.
If we place the setup for the tests in the test/1 or describe/1
blocks, then we can ensure that we can add a new describe block to
this test module with a fresh context.