Filter 'bbc-mvt' headers in request when creating request hash
We would like to vary on MVT request headers at the RouteState level.
Now that we have access to the MVT headers in :mvt_seen in the RouteState
state, we can vary the request hash by removing all MVT request
raw_headers that have experiment/feature names not found in :mvt_seen.