create world service topics redirect transformer
This transformer uses the path to decide if it should change the origin and platform to mozart. If the path contains a guid such as '/portuguese/topics/15f1bcf6-b6ab-48e8-b708-efed41e43d31' it should be sent to mozart otherwise it should still be served by simorgh. Non guid routes will have an id that is exactly 12 characters long