Move ResponseTransformer.MvtMapper to post response pipeline
Responses that are served from the cache do not pass through
the response pipeline - therefore the ResponseTransformer.MvtMapper
transformer will not be called, the request MVT raw_headers will not be
filtered against MVT vary headers in the response, and as a result all
of the MVT request headers will be present in the vary header returned
by Belfrage.

By moving ResponseTransformer.MvtMapper to the post response pipeline,
we ensure that it is always called - whether the response is returned from
the cache or the origin.