[RESFRAME-4702] Add consistency to MVT vary headers
If MVT experiment/feature names are present in the:

* Request headers [1]
* Vary header in the origin response [2]
* Slots config [3]

The MVT headers will be appended to the response header that we send from belfrage to the client, for example:
```
vary: Accept-Encoding,X-BBC-Edge-Cache,X-Country,X-IP_Is_UK_Combined,X-BBC-Edge-Scheme,bbc-mvt-1
```

We need to make a change so that if the experiment/feature name is not present in the request, but is present in the vary header in the origin response and the slots config, then the MVT headers will still be appended.

In order to do this we can update the logic in `Belfrage.Mvt.Mapper`. This module creates a Map in `struct.private.mvt` that contains parsed MVT experiment/feature names and values if they are present in the

* Request headers
* Slots config

for example:
```
%{"mvt-button_colour" => {1, "experiment;red"},
  "mvt-sidebar" => {3, "feature;false"}
  } = struct.private.mvt
```

We can change the logic so that if the experiment/feature names are present in the slots config but not the request headers, the second tuple element will be nil. For example if `button_colour` is not present in any MVT request header, the above will now be:
```
%{"mvt-button_colour" => {1, nil},
  "mvt-sidebar" => {3, "feature;false"}
  } = struct.private.mvt
```

`Belfrage.ResponseTransformer.MvtMapper` filters MVT headers from the allowlist that are not in `struct.private.mvt`. Since we now include MVT experiment/feature names that are only present in the slots config in `struct.private.mvt`, they will remain in the allowlist after the said `Transformer` is called. This is important as the allowlist is used to construct the vary header sent in the response from Belfrage.

---

[1] Example of an experiment/feature name present in the request:
```
bbc-mvt-1: experiment;sfv_verification_test;b
```

[2] Example of an experiment/feature name present in the vary header in the response from the origin:
```
vary: mvt-sfv_verification_test
```

[3] Example of an experiment/feature name present in the slots config:
```
{
  "projects": {
    "1": [
      {
        "key": "sfv_verification_test",
        "header": "bbc-mvt-1",
        ...
      }
    ]
  }
}
```