Fetch Slots config on startup of Mvt.FilePoller
In order to return an MVT header from Belfrage in the vary, we need
to have the MVT experiment/feature in the:

* Slots config
* Request headers
* Vary header in the origin response

The Slots config is fetched from the `Mvt.FilePoller`, which has a
delay of `60_000` before the first fetch is attempted. We need to
ensure that the first fetch is attempted on startup of the `Mvt.FilePoller`
`GenServer`.