Add Etag support on the RouteSpec level
This change enables us to support Etags on a `RouteSpec` level.

This also fixes the following bug:

Response body compression occurs in the `PreCacheCompression` `ResponseTranformer`,
and subsequent decompression occurs in the `CompressionAsRequested` `ResponseTranformer`.
Compression and subsequent descompression of a blank string results in the following zlib error:

    iex(2)> :zlib.unzip(:zlib.gzip(""))
    ** (ErlangError) Erlang error: :data_error
        :zlib.inflate_nif(#Reference<0.3595097228.3051749379.191595>, 8192, 16384, 0)
        :zlib.dequeue_all_chunks_1/3
        :zlib.inflate/3
        :zlib.unzip/1

To handle this, we should not attempt to compress a blank string.