Check cache status header in non-personalised cached response test
When a fresh cache response is served, in `Belfrage.Cache.Fetch`
`struct.private.origin` is set to `:belfrage_cache`.

In `BelfrageWeb.Response.Headers.CacheStatus` this should result in
the response header:

    belfrage-cache-status: HIT

being set.

This also holds for the scenario in which a non-personalised
response has been cached and a subsequent personalised request is
made and fails. A fresh cache response will be served with the
above response header.

This commit adds a check for the above.