Use new fallback: true opt. in Cache.fetch/3 when fetching fallback
For a given route, if a non-personalised response has been cached
and a subsequent personalised request is made and fails, the fresh
non-personalised cache response will be served.

Given that the fresh response in this scenario is in essence a
fallback, we should treat it as such, so that the
"belfrage-cache-status" response header is set to "STALE".

This is achieved by setting `struct.response.fallback` to `true`
when a fresh response is fetched from the cache and a `fallback: true`
option is given to `Cache.fetch/3`.

We then use the `fallback: true` option when fetching a fallback
in `Belfrage.Processor`.