Created `Expectations` module which contains helper to test response
In order to create a retry function we must stop using assertions and move all
testing logic into one place. As when we use assertions as soon as one fails
the test fails.

Instead we must use functions which validate the response which return an
meaningful error describing why the response isn't valid.

These errors can be passed through as a custom error to ExUnit so it can be
displayed in the case of a failing smoke test.