should now run the scheduled smoke tests at 8:30am
Changes the order of 'choice' parameters.
When a jenkins pipeline is first run you can't choose the parameters that it
runs with, it just uses the defualt. I this case it means that the smoke ttests
fire at test and hit all stacks. This wastes time and causes the smoke tests to
fail.

The 'choice' parameter doesn't use a default option, it just chooses the first
option. So I changed the first option to be ENVIRONMENT 'live' and TARGET
'belfrage'.

Moved an hour back so it will show results at 8:30am, for some reason the cron
job is in UTC when the rest of jenkins is in BST