[RESFRAME-4913] Return internal 404 response for invalid querystrings
Search terms that are invalid strings result in in 500 response.
This adds an end-to-end test for the above scenario and ensures
that a 404 is returned.

To test this the Belfrage.Lambda and Belfrage.AWS mocks
had to be removed, and the respective values have be retrieved
at runtime in order to get the updated values.