Introduces Belfrage.Xray.Enable which enables/disables xray in the processor.
If `xray_enabled` is true, it will do nothing.
If `xray_enabled` is false, it will set `xray_segment` to nil

This code has to be implemented here as `xray_enabled` defaults to false in routespecs.
This means that by default we will be removing `xray_segment`.
If this were to be a transformer is would have to be added to every platform we
want to DISABLE xray on.