RESFRAME-4941/simorgh-mvt (#1667)
* moved put_mvt_headers into separate module in order to share with http service

* created remove_original_headers/1

This is to remove redundant 'bbc-mvt-i' headers placed just after the Mvt.Mapper.map/1, so that they dont get sent downstream by the http service.

Note:
Today I discovered that when you filter a map it returns a list of tuples which
then have to be transformed back into a map again.

* Included Mvt.Headers.put_mvt_headers into the http service

Because put_mvt_headers/2 needs access to struct.private.mvt we need access to the whole struct this means:
 - build_headers/1 takes a struct NOT a struct.response
 - build_request(struct.request, struct.private.origin) -> build_request(struct)

* added tests to http service to show private.mvt contents getting added to headers

* added mvt end to end tests for http endpoints

* Addressing Sam's comments

- about put_mvt_headers/2 function description
- about putting mvt_project_id to mock routespec for now

* addressed davids suggestions

- typo not -> no
- using :maps.filter/2 instead of Enum.filter/2

* Addressed Joe comment bbc-mvt-i -> bbc-mvt-{id}

* Removed alias which caused compiler warnings

* addressed sams comment of joes comment replaced bbc-mvt-{id} with bbc-mvt-{i}

* removes mvt- override headers on test but not on live

* mvt end to end test also checks for recieved headers

* changed remove_original_headers to not remove bbc-mvt-complete

* mvt mapper doesn't filter 'bbc-mvt-complete' from the allowlist if present in allowlist and vary header

* added 'bbc-mvt-complete' header into end to end tests, they are now passing

* small neatening of mvt_mapper

* implemented own starts_with?/2 function which compares string with string, rather than string with list

* 'bbc-mvt-complete' is filtered from raw_headers and will no longer be put in request headers to the origin

* 'bbc-mvt-complete' header is now placed into the struct.private.mvt

* 'bbc-mvt-complete' header is now added to vary header when present in struct.private.mvt

* end to end tests now pass with correct 'bbc-mvt-complete' functionality

* addressed sams comment

github discussion: https://github.com/bbc/belfrage/pull/1667\#discussion_r1011377377

* addressed typos

* Updated behaviour of 'bbc-mvt-complete' with the vary header ...

'bbc-mvt-complete' header is now always part of the vary header IF the route
the request belongs to have MVT enabled.

* removed mvt_complete/1 function as no longer neccessary

as per: https://github.com/bbc/belfrage/pull/1667\#discussion_r1016996353

* added comment to  test

as per: https://github.com/bbc/belfrage/pull/1667\#discussion_r1016999412

* changed test message from 'headers' to 'raw_headers'

as per: https://github.com/bbc/belfrage/pull/1667\#discussion_r1017000406

* Allow mvt_mapper to pass '1' or '0' as values for 'bbc-mvt-complete' header

The bbc-mvt-complete headers purpose is to let services that use the MVT
service if the response that it gives you can be trusted.

If the value is '1' then the MVT service is saying you can trust what we are
giving you. As the service is fully operational.

If the value is a '0' then the response might not return reliable data because
MVT service is degraded. Maybe an S3 bucket is down, so the data being given
may be out of date.

* mvt_mapper (the response transformer) now runs if  if private.mvt is NOT empty

before this change it added mvt complete if the vary header (from the origin)
did NOT have any mvt headers.

* modifies test to show 'bbc-mvt-complete' included in vary header even if response cached

* added test to show bbc-mvt-complete with value {nil, '0'} can be parsed