Fix handler detachment bug in Belfrage.Xray.Telemetry
The following error is logged when `mix test` is run:

    {"route_state_id":"SomeRouteState","metadata":{},"message":"Handler \"belfrage-xray\" has failed and has been detached. Class=:error\nReason=:undef\nStacktrace=[\n  {nil, :trace, [], []},\n  {Belfrage.Xray.Telemetry, :subsegment, 5,\n   [file: 'lib/belfrage/xray/xray_telemetry.ex', line: 26]},\n  {Belfrage.Xray.Telemetry, :handle_event, 4,\n   [file: 'lib/belfrage/xray/xray_telemetry.ex', line: 18]},\n  {:telemetry, :\"-execute/3-fun-0-\", 4,\n

This happens because when the event metadata does not contain
a segment - specifically, metadata.struct.request.xray_segment
is `nil`. This means that in Belfrage.XrayTelemetry.subsegment/4
when an attempt to access segment.trace.sampled is made, an error
will be thrown and as per the documentation for
:telemetry.attach_many/4 [1] the handler will
be detached for further events.

This also fixes an xray end-to-end test.

[1] https://hexdocs.pm/telemetry/1.0.0/telemetry.html#attach_many/4