Add Belfrage.Metrics.NimblePool.properties/1
Finch executes the following events [1]:

> ### Queue Start
> `[:finch, :queue, :start]` - Executed before checking out a connection from the pool.
> #### Measurements
>   * `:system_time` - The system time.
> #### Metadata
>   * `:pool` - The pool's PID.
>   * `:request` - The request (`Finch.Request`).
> ### Queue Stop
> `[:finch, :queue, :stop]` - Executed after a connection is retrieved from the pool.
> #### Measurements
>   * `:duration` - Time taken to check out a pool connection.
>   * `:idle_time` - Elapsed time since the connection was last checked in or initialized.
> #### Metadata
>   * `:pool` - The pool's PID.
>   * `:request` - The request (`Finch.Request`).

The metadata contains the pool's PID, but not the host that the
pool is mapped to.

We would like to tag any metrics from these events with the
host for the pool, as these are what we are using to identify
our pools.

To this end we add a public properties/1 function that
contains pool properties, including the host for the pool.

[1] https://hexdocs.pm/finch/0.13.0/Finch.Telemetry.html