Fix Belfrage.BelfrageCacheTest tests
The module attribute @cache_seeded_response contains a :cache_last_updated
field set to the value of Belfrage.Timer.now_ms(). This will be set
when the module is compiled, and will not be set dynamically when
the test is run, which can lead to unexpected test failures.

Instead, the :cache_last_updated value is set in the unit test.