Fix platform response  double metricing and remove broken belfrage webcore response (#2341)
* Remove platform response without route_spec

Since we call platform.response with `route_spec`, `platform`, `status_code`, it looks like it will also hit the clause where it looks for no `route_spec`. The desired outcome was for this to catch platform.response when `route_spec` was not available however this will take some more thought. We may experience some nils in RouteSpec for Cloudwatch metrics however I think the dimension may get dropped at this point.
This won't effect metrics reporting as we should still this metric available because of Cloudwatch aggregation giving us the metric with and without the routespec.

* Remove belfrage.webcore.response

We don't use this metric in the graph, we use the platform.response metric.
This has also had a typo in it for a long time so we haven't reported the metric for a long time either.